backend/requirements.txt
django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
requests==2.32.3
psycopg2-binary==2.9.9
gunicorn==23.0.0
PyJWT==2.9.0
cryptography==43.0.1